Add another kong for internal ingress controller
This is used to route requests for Jupyter Notebooks.

The original Kong receives a request for "/v1/sessions" and
routes it to Session Manager Server/Agent. Then Session
Manager Agent routes the request to this internal Kong with
the same URL ("/v1/sessions").

Originally we are planning to make Session Manager truncate
"/v1/sessions/<cluster ID>", but that didn't work. Jupyter Notebook
didn't work when its ServerApp.base_url is different from the
user visible path.